requirements.txt
alembic==1.17.1
annotated-types==0.7.0
black==25.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
greenlet==3.3.0
idna==3.11
isort==7.0.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.1
marshmallow-sqlalchemy==1.4.2
mypy_extensions==1.1.0
nats-py==2.12.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.3.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
